Propagate error instead of os.Exit in goroutine
Calling os.Exit() is unfounded for a function that is meant to be consumed as a module in other Go code, but addLicense's historical behavior is to return a non-zero exit code when the `-check` flag is used and headers are missing from files.

To get around this, I updated the goroutine/closure setup to update a higher-scoped variable with any error message after `wg.Wait()` is called. This effectively lets us track if an error has occurred during the goroutine that walks all folders/files.

That error is then propagated upward by the `Run()` function. To retain historical exit codes, I modified the `main` CLI function to check specifically for the `missing license header` case and call `os.Exit(1)` there.